# src/api/rest/docs/swagger.yaml
basePath: /tumblebug
definitions:
...
  model.BastionNode:
    properties:
      mciId:
        type: string
      vmId:
        type: string
    type: object
...
  model.IdList:
    properties:
      output:
        items:
          type: string
        type: array
    type: object
...
  model.KeyValue:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        example: Any message
        type: string
    type: object
...
  model.TbSshKeyInfo:
    properties:
      associatedObjectList:
        items:
          type: string
        type: array
      connectionName:
        type: string
      cspSshKeyId:
        description: CspSshKeyId used for CSP-native identifier (either Name or ID)
        type: string
      cspSshKeyName:
        description: CspSshKeyName used for CB-Spider identifier
        type: string
      description:
        type: string
      fingerprint:
        type: string
      id:
        type: string
      isAutoGenerated:
        type: boolean
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        type: string
      privateKey:
        type: string
      publicKey:
        type: string
      systemLabel:
        description: SystemLabel is for describing the Resource in a keyword (any
          string can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      username:
        type: string
      uuid:
        description: uuid is universally unique identifier for the resource
        type: string
      verifiedUsername:
        type: string
    type: object
  model.TbSshKeyReq:
    properties:
      connectionName:
        type: string
      cspSshKeyId:
        description: |-
          Fields for "Register existing SSH keys" feature
          CspSshKeyId is required to register object from CSP (option=register)
        type: string
      description:
        type: string
      fingerprint:
        type: string
      name:
        type: string
      privateKey:
        type: string
      publicKey:
        type: string
      username:
        type: string
      verifiedUsername:
        type: string
    required:
    - connectionName
    - name
    type: object
  model.TbSubnetInfo:
    properties:
      bastionNodes:
        items:
          $ref: '#/definitions/model.BastionNode'
        type: array
      description:
        type: string
      id:
        type: string
      idFromCsp:
        type: string
      ipv4_CIDR:
        type: string
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        type: string
      uuid:
        description: uuid is universally unique identifier for the resource
        type: string
    required:
    - ipv4_CIDR
    - name
    type: object
  model.TbSubnetReq:
    properties:
      description:
        type: string
      idFromCsp:
        type: string
      ipv4_CIDR:
        type: string
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        type: string
    required:
    - ipv4_CIDR
    - name
    type: object
...
  resource.JSONResult:
    type: object
...
  resource.RestGetAllSshKeyResponse:
    properties:
      sshKey:
        items:
          $ref: '#/definitions/model.TbSshKeyInfo'
        type: array
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cb-tumblebug/issues/new/choose
  description: CB-Tumblebug is an open source system for managing multi-cloud infrastructure
    consisting of resources from multiple cloud service providers. (Cloud-Barista)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/cloud-barista/cb-tumblebug/blob/main/README.md
  title: CB-Tumblebug REST API
  version: latest
paths:
...
  /ns/{nsId}/resources/sshKey:
    delete:
      consumes:
      - application/json
      description: Delete all SSH Keys
      operationId: DelAllSshKey
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: ""
        description: Delete resources containing matched ID-substring only
        in: query
        name: match
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IdList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete all SSH Keys
      tags:
      - '[Infra Resource] Access Key Management'
    get:
      consumes:
      - application/json
      description: List all SSH Keys or SSH Keys' ID
      operationId: GetAllSshKey
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Option
        enum:
        - id
        in: query
        name: option
        type: string
      - description: 'Field key for filtering (ex: systemLabel)'
        in: query
        name: filterKey
        type: string
      - description: 'Field value for filtering (ex: Registered from CSP resource)'
        in: query
        name: filterVal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given option param
          schema:
            allOf:
            - $ref: '#/definitions/resource.JSONResult'
            - properties:
                '[DEFAULT]':
                  $ref: '#/definitions/resource.RestGetAllSshKeyResponse'
                '[ID]':
                  $ref: '#/definitions/model.IdList'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List all SSH Keys or SSH Keys' ID
      tags:
      - '[Infra Resource] Access Key Management'
    post:
      consumes:
      - application/json
      description: Create SSH Key
      operationId: PostSshKey
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: 'Option: [required params for register] connectionName, name,
          cspSshKeyId, fingerprint, username, publicKey, privateKey'
        enum:
        - register
        in: query
        name: option
        type: string
      - description: Details for an SSH Key object
        in: body
        name: sshKeyInfo
        required: true
        schema:
          $ref: '#/definitions/model.TbSshKeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSshKeyInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create SSH Key
      tags:
      - '[Infra Resource] Access Key Management'
  /ns/{nsId}/resources/sshKey/{sshKeyId}:
    delete:
      consumes:
      - application/json
      description: Delete SSH Key
      operationId: DelSshKey
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: SSH Key ID
        in: path
        name: sshKeyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete SSH Key
      tags:
      - '[Infra Resource] Access Key Management'
    get:
      consumes:
      - application/json
      description: Get SSH Key
      operationId: GetSshKey
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: SSH Key ID
        in: path
        name: sshKeyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSshKeyInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get SSH Key
      tags:
      - '[Infra Resource] Access Key Management'
    put:
      consumes:
      - application/json
      description: Update SSH Key
      operationId: PutSshKey
      parameters:
      - description: Details for an SSH Key object
        in: body
        name: sshKeyInfo
        required: true
        schema:
          $ref: '#/definitions/model.TbSshKeyInfo'
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: SshKey ID
        in: path
        name: sshKeyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSshKeyInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Update SSH Key
      tags:
      - '[Infra Resource] Access Key Management'
...
  /ns/{nsId}/resources/vNet/{vNetId}/subnet:
    post:
      consumes:
      - application/json
      description: Create Subnet
      operationId: PostSubnet
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: VNet ID
        in: path
        name: vNetId
        required: true
        type: string
      - description: Details for an Subnet object
        in: body
        name: subnetReq
        required: true
        schema:
          $ref: '#/definitions/model.TbSubnetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSubnetInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create Subnet
      tags:
      - '[Infra Resource] Network Management'
  /ns/{nsId}/resources/vNet/{vNetId}/subnet/{subnetId}:
    delete:
      consumes:
      - application/json
      description: Delete Subnet
      operationId: DelSubnet
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: VNet ID
        in: path
        name: vNetId
        required: true
        type: string
      - description: Subnet ID
        in: path
        name: subnetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete Subnet
      tags:
      - '[Infra Resource] Network Management'
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token ([TBD] Get token
      in http://xxx.xxx.xxx.xxx:xxx/auth)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"