# src/core/model/vnet.go
/*
Copyright 2019 The Cloud-Barista Authors.
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
    http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

// Package model is to handle object of CB-Tumblebug
package model

// 2020-04-09 https://github.com/cloud-barista/cb-spider/blob/master/cloud-control-manager/cloud-driver/interfaces/resources/VPCHandler.go

// SpiderVPCReqInfoWrapper is a wrapper struct to create JSON body of 'Create VPC request'
type SpiderVPCReqInfoWrapper struct {
	ConnectionName string
	ReqInfo        SpiderVPCReqInfo
}

// SpiderVPCReqInfo is a struct to create JSON body of 'Create VPC request'
type SpiderVPCReqInfo struct {
	Name           string
	IPv4_CIDR      string
	SubnetInfoList []SpiderSubnetReqInfo
	CSPId          string
}

// SpiderVPCInfo is a struct to handle VPC information from the CB-Spider's REST API response
type SpiderVPCInfo struct {
	IId            IID // {NameId, SystemId}
	IPv4_CIDR      string
	SubnetInfoList []SpiderSubnetInfo
	KeyValueList   []KeyValue
}

// TbVNetReq is a struct to handle 'Create vNet' request toward CB-Tumblebug.
type TbVNetReq struct { // Tumblebug
	Name           string        `json:"name" validate:"required"`
	ConnectionName string        `json:"connectionName" validate:"required"`
	CidrBlock      string        `json:"cidrBlock"`
	SubnetInfoList []TbSubnetReq `json:"subnetInfoList"`
	Description    string        `json:"description"`
	CspVNetId      string        `json:"cspVNetId"`
}

// TbVNetInfo is a struct that represents TB vNet object.
type TbVNetInfo struct { // Tumblebug
	Id   string `json:"id"`
	Name string `json:"name"`
	// uuid is universally unique identifier for the resource
	Uuid                 string         `json:"uuid,omitempty"`
	ConnectionName       string         `json:"connectionName"`
	CidrBlock            string         `json:"cidrBlock"`
	SubnetInfoList       []TbSubnetInfo `json:"subnetInfoList"`
	Description          string         `json:"description"`
	CspVNetId            string         `json:"cspVNetId"`
	CspVNetName          string         `json:"cspVNetName"`
	Status               string         `json:"status"`
	KeyValueList         []KeyValue     `json:"keyValueList"`
	AssociatedObjectList []string       `json:"associatedObjectList"`
	IsAutoGenerated      bool           `json:"isAutoGenerated"`

	// SystemLabel is for describing the Resource in a keyword (any string can be used) for special System purpose
	SystemLabel string `json:"systemLabel" example:"Managed by CB-Tumblebug" default:""`

	// Disabled for now
	//Region         string `json:"region"`
	//ResourceGroupName string `json:"resourceGroupName"`
}

// BastionNode is a struct that represents TB BastionNode object.
type BastionNode struct {
	MciId string `json:"mciId"`
	VmId  string `json:"vmId"`
}